+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Added
 - Added functionality for the agent `--allow-list` configuration, which
 whitelists check and check hook executables.
+- Added backwards compatible content negotiation to the websocket connection.
+Protobuf will be used for serialization/deserialization unless indicated by the
+backend to use JSON.
 
 ### Fixed
 - Fixed the tabular output of `sensuctl filter list` so inclusive filter expressions
@@ -22,8 +25,6 @@ queried in the `GET /events/:entity` endpoint (#3141)
 ### Fixed
 - The agent now sends heartbeats to the backend in order to detect network
 failures and reconnect faster.
-- Added the `--protobuf-serialization` flag to the agent to use protobuf for
-serialization/deserialization rather than the default, JSON.
 
 ### Changed
 - The default handshake timeout for the WebSocket connection negotiation has